Address Barrett's comments
